moved exceptions handler to separate folder
